second commit

+++ database/migrations/2023_07_03_001456_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->bigIncrements('id');    
             $table->string('nama_barang');
-            $table->integer('harga');
+            $table->float('harga', 12, 4);
             $table->integer('stok');
             $table->string('foto');
             $table->timestamps();
